slab_free() bug fix, slab_destroy is broken
slab_free() didn't move a partial slab to the free state if all objects
were free, this commit fixes said issue.

slab_destroy is currently broken, the cache_list is not setting prev and
next pointers correctly but I'm too tired to fix it right now.
'slab_caches' has also been turned into a 'struct cache_list' with head
and tail members so that when the list is traversed, we still have a
reference to the head and tail of the cache list.